update documentation and refactoring mapper
