+++ abstracts/__init__.py
@@ -1,2 +1,1 @@
 from .base_adaptor import Adaptor
-from .container_orchestrator import ContainerAdaptor
