--- abstracts/cloudorchestrator.py
@@ -1,6 +1,0 @@
-from . import base_adaptor
-
-class CloudAdaptor(base_adaptor.Adaptor):
-
-    def __init__(self):
-        super(CloudAdaptor, self).__init__()
