--- abstracts/container_orchestrator.py
@@ -1,6 +1,0 @@
-from . import base_adaptor
-
-class ContainerAdaptor(base_adaptor.Adaptor):
-
-    def __init__(self):
-        super(ContainerAdaptor, self).__init__()
