--- abstracts/policykeeper.py
@@ -1,6 +1,0 @@
-from . import base_adaptor
-
-class PolicyKeeperAdaptor(base_adaptor.Adaptor):
-
-    def __init__(self):
-        super(PolicyKeeperAdaptor, self).__init__()
