--- abstracts/securityenforcer.py
@@ -1,6 +1,0 @@
-from . import base_adaptor
-
-class SecurityEnforcerAdaptor(base_adaptor.Adaptor):
-
-    def __init__(self):
-        super(SecurityEnforcerAdaptor, self).__init__()
