+++ documentation/source/component_submitter.abstracts.rst
@@ -12,42 +12,10 @@ component\_submitter.abstracts.base\_adaptor module
     :undoc-members:
     :show-inheritance:
 
-component\_submitter.abstracts.cloudorchestrator module
--------------------------------------------------------
-
-.. automodule:: component_submitter.abstracts.cloudorchestrator
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-component\_submitter.abstracts.container\_orchestrator module
--------------------------------------------------------------
-
-.. automodule:: component_submitter.abstracts.container_orchestrator
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
 component\_submitter.abstracts.exceptions module
 ------------------------------------------------
 
 .. automodule:: component_submitter.abstracts.exceptions
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-component\_submitter.abstracts.policykeeper module
---------------------------------------------------
-
-.. automodule:: component_submitter.abstracts.policykeeper
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-component\_submitter.abstracts.securityenforcer module
-------------------------------------------------------
-
-.. automodule:: component_submitter.abstracts.securityenforcer
     :members:
     :undoc-members:
     :show-inheritance:
