+++ documentation/source/component_submitter.tests.rst
@@ -1,0 +1,30 @@
+component\_submitter.tests package
+==================================
+
+Submodules
+----------
+
+component\_submitter.tests.test\_docker module
+----------------------------------------------
+
+.. automodule:: component_submitter.tests.test_docker
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+component\_submitter.tests.test\_validator module
+-------------------------------------------------
+
+.. automodule:: component_submitter.tests.test_validator
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: component_submitter.tests
+    :members:
+    :undoc-members:
+    :show-inheritance:
