+++ system/key_config.yml
@@ -1,5 +1,5 @@
-config:
-  dry_run: False
+main_config:
+  dry_run: True
   #path_log: "/var/lib/toscasubmitter/submitter.log"
   path_log: "submitter.log"
 step:
@@ -35,11 +35,11 @@ adaptor_config:
      - "tosca.policies.Scaling.*"
    endoint: "endpoint"
    volume: "/var/lib/submitter/security_workdir_example/"
-
  DockerAdaptor:
    types:
      - "tosca.nodes.MiCADO.Container.Application.Docker"
    endoint: "endpoint"
+
    volume: "/var/lib/submitter/files/output_configs/"
